requirements.txt
Flask
cs50
flask_session
requests
pycryptodome